Add UT and align Maven/MIMA to 3.9.10
