+++ pom.xml
@@ -71,9 +71,9 @@
 
     <!-- Dependency versions -->
     <version.maveniverseShared>0.1.8</version.maveniverseShared>
-    <version.maven>3.9.9</version.maven>
+    <version.maven>3.9.10</version.maven>
     <version.resolver>1.9.23</version.resolver>
-    <version.mima>2.4.28</version.mima>
+    <version.mima>2.4.29</version.mima>
     <version.slf4j>1.7.36</version.slf4j>
   </properties>
